+++ core/evol/method/select/Selector.h
@@ -19,7 +19,7 @@ class Selector {
   /**
    * @brief Performs selection of `size` instances from population.
    */
-  virtual void select(instance::Population& population, size_t size) = 0;
+  virtual void select(instance::Population& population, uint32_t size) = 0;
 };
 
 using RSelector = std::shared_ptr<Selector>;
